pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "hatch-build-scripts", "ruff>=0.9.0"]
build-backend = "hatchling.build"

[project]
name = "zae-limiter"
dynamic = ["version"]
description = "Rate limiting library backed by DynamoDB with token bucket algorithm"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    { name = "ZeroAE", email = "dev@zeroae.com" },
]
keywords = ["rate-limiting", "dynamodb", "token-bucket", "aws", "llm", "api-rate-limit"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aioboto3>=12.0.0",
    "aws-lambda-builders>=1.40.0",
    "boto3>=1.34.0",
    "click>=8.0.0",
    "pip",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
]

[project.optional-dependencies]
dev = [
    "boto3-stubs[dynamodb,s3,lambda,cloudformation,sts]>=1.34.0",
    "cfn-lint>=1.43.3",
    "diff-cover>=9.0.0",
    "mypy>=1.8.0",
    "moto[dynamodb]>=5.0.0",
    "openapi-spec-validator>=0.7.1",  # Required by moto but not declared
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.0.0",
    "pytest-repeat>=0.9.4",
    "pytest-examples>=0.0.18",
    "pytest-xdist>=3.5.0",
    "ruff>=0.1.0",
    "types-aiobotocore[dynamodb,s3]>=2.0.0",
    "types-docker>=7.0.0",
    "types-gevent>=24.0.0",
    "types-PyYAML>=6.0",
]
lambda = [
    "aws-lambda-powertools>=2.0.0",
]
plot = [
    "asciichartpy>=1.5.25",
]
local = [
    "docker>=7.0.0",
]
bench = [
    "docker>=7.0.0",
    "locust>=2.20",
    "gevent>=23.0",
]
cdk = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-click>=0.8.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
    "mike>=2.1.0",
]

[project.urls]
Homepage = "https://github.com/zeroae/zae-limiter"
Documentation = "https://zeroae.github.io/zae-limiter/"
Repository = "https://github.com/zeroae/zae-limiter"

[project.scripts]
zae-limiter = "zae_limiter.cli:cli"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/zae_limiter/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/zae_limiter", "src/zae_limiter_aggregator", "src/zae_limiter_provisioner"]

[[tool.hatch.build.hooks.build-scripts.scripts]]
commands = ["python scripts/generate_sync.py"]

[tool.hatch.envs.default.scripts]
generate-sync = "python scripts/generate_sync.py"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short -n auto --dist loadscope --import-mode=importlib --benchmark-disable"
markers = [
    "integration: marks tests as integration tests (require Docker or AWS)",
    "benchmark: marks tests as performance benchmarks",
    "aws: marks tests as real AWS tests (require --run-aws flag)",
    "e2e: marks tests as end-to-end tests (full stack lifecycle)",
    "slow: marks tests with significant wait times (>30s sleeps)",
    "monitoring: marks tests that verify CloudWatch alarms, metrics, DLQ",
    "snapshots: marks tests that verify usage snapshot creation",
    "doctest: marks tests that validate documentation code examples",
    "gevent: marks tests that trigger gevent monkey-patching (incompatible with xdist)",
]
filterwarnings = [
    "ignore:StackOptions is deprecated:DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Generated sync files: ast.unparse doesn't preserve line breaks in strings
"src/zae_limiter/sync_*.py" = ["E501"]
"src/zae_limiter/infra/sync_*.py" = ["E501"]
"tests/unit/test_sync_*.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = [
    "aws_lambda_builders.*",
    "aioboto3",
    "aioboto3.*",
    "ulid",
    "zae_limiter._version",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["locust.*", "locustfile"]
ignore_missing_imports = true

[tool.coverage.run]
omit = [
    # Generated sync files - coverage measured on async source only
    "*/sync_*.py",
]

